Added webserver dockerfile and build/run scripts
Added a dockerfile that builds an image from the node:alpine
image (a lightweight linux distribution with node preinstalled),
then installs typescript and the npm dependencies, then copies
our app over, compiles it, and runs it.

Added a .dockerignore file, which tells the docker COPY command
to ignore certain files (much like .gitignore).

Added a docker_build.sh script, which removes the existing container
(i.e. the previous run of the webserver; this is necessary in order to
remove the previous build of the image), removes the old image (and
no-longer-necessary intermediate layers), and builds a new
image tagged as tindart/webserver:latest.

Added a docker_run.sh script, which removes the existing container
and runs a new instance of the webserver image.